+++ python/capitulo01/ex1.3.py
@@ -1,6 +1,6 @@
-numero_1 = int(input("Número 1: "))
-numero_2 = int(input("Número 2: "))
-numero_3 = int(input("Número 3: "))
+numero_1 = float(input("Número 1: "))
+numero_2 = float(input("Número 2: "))
+numero_3 = float(input("Número 3: "))
 
 media_aritmetica = (numero_1 + numero_2 + numero_3) /3
